requirements.txt
# python-telegram-bot v21+ hỗ trợ Python 3.13
python-telegram-bot[webhooks]>=21.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow>=10.4.0
